Struct cc_fs::fs::CcFs

struct CcFs {
    index: Index,
    tar: File,
    next_file_handle: u64,
}

FUSE file system with integrity protection backed by a tar file.

Fields
index: Index

Index for the tar file.

tar: File

Tar file backing store for the layer.

next_file_handle: u64

The next available file handle.

Implementations

impl CcFs

pub fn new(index: &String, tar: &String) -> Result<CcFs>

Create a new CcFs instance backed by a tar file.

Arguments

index - The index file to use for enforcing integrity.

tar - The tar file to use for file content backing store.

fn to_file_type(typeflag: &FileType) -> FileType

Map from CcFs FileType to FUSE FileType.

Arguments

typeflag - The CcFs FileType of the inode.

fn inode_to_attr(ino: u64, inode: &Inode) -> FileAttr

Fetch FUSE attributes for an inode.

Arguments

ino - Number of the inode.

inode - The inode.