Module cc_fs::index

Representation of indexes of confidential container file-systems.

An Index contains a set of Inodes representing the entries in a file-system, and also the hash states for veryfying the integrity of the file-system. The contents of each regular file in the file-system resides outside the index and can be provided from original tar files of container layers or from files in regular file-systems.

An Index is optimized both for lookup as well as memory consumption. The inodes are maintained in a sorted vec ordered by nesting depth, parent path, and name. Each directory inode also holds the position of its first child in the vec, and the number of children.

Indexes are serialized/deserialized using bincode which is a fast, compact binary format. Due to use of serde derive, use of many other formats (cbor, messagepack, postcard, json) is possible.

Structs

Extra

Infrequent properties of an item. Usually specified using PAX extensions.

Index

Index of a confidential container file-system.

Inode

Index node (Inode) of an item in file system.

Enums

FileType

Type of an item in the file-system.