Struct cc_fs::index::Inode

pub struct Inode {

    pub typeflag: FileType,
    pub name: String,
    pub parent: String,
    pub size: u32,
    pub uid: u32,
    pub gid: u32,
    pub mode: u32,
    pub mtime: u64,
    pub extra: Option<Extra>,
    pub num: u32,
    pub hash_index: u32,
    pub child_inode: u32,
    pub num_children: u32,
    pub offset: u32,
    pub depth: u16,
    pub links: u16,
    pub target_ino: u32,

}

Index node (Inode) of an item in file system.

Fields
typeflag: FileType

File type.

name: String

Name of the item.

parent: String

Path of the directory containing the item. The parent path must begin and end with ‘/’.

size: u32uid: u32gid: u32mode: u32mtime: u64extra: Option<Extra>

Infrequently occuring properties.

num: u32

The inode number of this inode.

hash_index: u32

Index of starting hash state.

child_inode: u32

Inode number of first child.

num_children: u32

Number of (direct) children.

offset: u32

512-block offset of the file in the backing tar file. Meaningful only for regular files.

depth: u16

The nesting level of this inode.

links: u16

Number of hard links to this inode.

target_ino: u32

Inode number of hard-link target.

Implementations

impl Inode

Implementation.

pub fn path_eq(&self, path: &String) -> bool

Check whether the inode has given path.